Point out important usage of nested Navigators
When developing the flow for the creation of one of many configurable objects of the same type, e.g. an item in a todo list, I ran into a road block creating multiple routes that could access this item. The reason being that every time I pushed a new route onto the root `Navigator`, it would build the widget as a child of the root `Navigator`. This made it so that I couldn't access the item using `BuildContext.inheritFromWidgetOfExactType` any more.